requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.18.0
et_xmlfile==2.0.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-anthropic==1.4.1
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-protocol==0.0.12
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.36
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.20.0
numpy==2.4.4
openpyxl==3.1.5
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
pandas==3.0.2
pillow==12.2.0
propcache==0.4.1
protobuf==7.34.1
pyarrow==24.0.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
streamlit==1.56.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
